完成DOA接口设计和Service层接口设计

+++ src/main/java/com/nchu/dao/AfterSaleDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.AfterSale;
  * 2017年9月20日09:28:03
  * 售后表Dao接口
  */
-public interface AfterSaleDao extends BaseDao<AfterSale,Integer>{
+public interface AfterSaleDao extends BaseDao<AfterSale,Long>{
 }
