+++ src/main/java/com/nchu/dao/AnnouncementDao.java
@@ -6,6 +6,6 @@ import com.nchu.entity.Announcement;
  * 2017年9月20日09:24:35
  * 公告表DAO接口
  */
-public interface AnnouncementDao extends BaseDao<Announcement,Integer> {
+public interface AnnouncementDao extends BaseDao<Announcement,Long> {
 
 }
