+++ src/main/java/com/nchu/dao/EvaluationDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.Evaluation;
  * 2017年9月20日10:03:54
  * 评论表DAO接口
  */
-public interface EvaluationDao extends BaseDao<Evaluation,Integer> {
+public interface EvaluationDao extends BaseDao<Evaluation,Long> {
 }
