+++ src/main/java/com/nchu/dao/ExpressDeliveryDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.ExpressDelivery;
  * 2017年9月20日10:14:25
  * 快递表DAO接口
  */
-public interface ExpressDeliveryDao extends BaseDao<ExpressDelivery,Integer> {
+public interface ExpressDeliveryDao extends BaseDao<ExpressDelivery,Long> {
 }
