+++ src/main/java/com/nchu/dao/FavoritesDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.Favorites;
  * 2017年9月20日09:29:38
  * 用户收藏夹表DAO接口
  */
-public interface FavoritesDao extends BaseDao<Favorites,Integer>{
+public interface FavoritesDao extends BaseDao<Favorites,Long>{
 }
