+++ src/main/java/com/nchu/dao/GoodsPictureDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.GoodsPicture;
  * 2017年9月20日14:03:54
  * 图片表DAO接口
  */
-public interface GoodsPictureDao extends BaseDao<GoodsPicture,Integer> {
+public interface GoodsPictureDao extends BaseDao<GoodsPicture,Long> {
 }
