+++ src/main/java/com/nchu/dao/GroupPurchaseDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.GroupPurchase;
  * 2017年9月20日10:09:30
  * 参团表Dao接口
  */
-public interface GroupPurchaseDao extends BaseDao<GroupPurchase,Integer> {
+public interface GroupPurchaseDao extends BaseDao<GroupPurchase,Long> {
 }
