+++ src/main/java/com/nchu/dao/OrderDao.java
@@ -6,6 +6,6 @@ import com.nchu.entity.Order;
  * 2017年9月20日09:21:32
  * 商品订单Dao接口
  */
-public interface OrderDao extends BaseDao<Order,Integer> {
+public interface OrderDao extends BaseDao<Order,Long> {
 
 }
