+++ src/main/java/com/nchu/dao/ReceiverAddressDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.ReceivingAddress;
  * 2017年9月20日13:59:26
  * 收货地址Dao接口
  */
-public interface ReceiverAddressDao extends BaseDao<ReceivingAddress,Integer>{
+public interface ReceiverAddressDao extends BaseDao<ReceivingAddress,Long>{
 }
