+++ src/main/java/com/nchu/dao/SaleTypeDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.Saletype;
  * 2017年9月20日15:24:16
  * 销售类型表的DAO接口
  */
-public interface SaleTypeDao extends BaseDao<Saletype, Integer> {
+public interface SaleTypeDao extends BaseDao<Saletype, Long> {
 }
