+++ src/main/java/com/nchu/dao/ShopDao.java
@@ -6,6 +6,6 @@ import com.nchu.entity.Shop;
  * 2017年9月20日08:12:55
  * 商店DAO接口
  */
-public interface ShopDao extends BaseDao<Shop,Integer> {
+public interface ShopDao extends BaseDao<Shop,Long> {
 
 }
