+++ src/main/java/com/nchu/dao/VouchersDao.java
@@ -6,5 +6,5 @@ import com.nchu.entity.Vouchers;
  * 2017年9月20日13:56:01
  * 优惠券Dao接口
  */
-public interface VouchersDao extends BaseDao<Vouchers,Integer>{
+public interface VouchersDao extends BaseDao<Vouchers,Long>{
 }
